Make homescreen widgets update daily
I was trying to write a Kotlin function to
trigger these homescreen widget updates,
but today I saw it didn't work on my phone.
I know these Dart workmanager calls work
daily, because the notifications are firing for
me at 8am, so I've swapped to using them
for triggering the widget updates.
I'll check in on this tomorrow and hopefully it's
worked.